// ClientApp/models/teamsite.ts
export interface TeamSite {
	date: string;
	day: string;
	sV1Name: string;
	sV1Site: string;
	drI1Name: string;
	drI1Site: string;
	drI2Name: string;
	drI2Site: string;
	rN1Name: string;
	rN1Site: string;
	rN2Name: string;
	rN2Site: string;
	rN3Name: string;
	rN3Site: string;
	ccA1Name: string;
	ccA1Site: string;
	ccA2Name: string;
	ccA2Site: string;
	ccA3Name: string;
	ccA3Site: string;
}