// ClientApp/models/team.ts
export interface Team {
	id: number;
	name: string;
	sV1Id: number;
	sV1Name: string;
	drI1Id: number;
	drI1Name: string;
	drI2Id: number;
	drI2Name: string;
	rN1Id: number;
	rN1Name: string;
	rN2Id: number;
	rN2Name: string;
	rN3Id: number;
	rN3Name: string;
	ccA1Id: number;
	ccA1Name: string;
	ccA2Id: number;
	ccA2Name: string;
	ccA3Id: number;
	ccA3Name: string;
}